Implement Derivers for encoders
