+++ modules/core/src/main/scala/me/mnedokushev/zio/apache/arrow/core/codec/ValueVectorCodec.scala
@@ -2,9 +2,9 @@ package me.mnedokushev.zio.apache.arrow.core.codec
 
 import org.apache.arrow.memory.BufferAllocator
 import org.apache.arrow.vector.ValueVector
-import org.apache.arrow.vector.complex.{ ListVector, StructVector }
+// import org.apache.arrow.vector.complex.{ ListVector, StructVector }
 import zio._
-import zio.schema.Schema
+// import zio.schema.Schema
 
 final case class ValueVectorCodec[A, V <: ValueVector](
   encoder: ValueVectorEncoder[A, V],
@@ -28,18 +28,18 @@ final case class ValueVectorCodec[A, V <: ValueVector](
 
 }
 
-object ValueVectorCodec {
+// object ValueVectorCodec {
 
-  def apply[A, V <: ValueVector](implicit codec: ValueVectorCodec[A, V]): ValueVectorCodec[A, V] =
-    codec
+//   def apply[A, V <: ValueVector](implicit codec: ValueVectorCodec[A, V]): ValueVectorCodec[A, V] =
+//     codec
 
-  implicit def primitive[A, V <: ValueVector](implicit schema: Schema[A]): ValueVectorCodec[A, V] =
-    ValueVectorCodec(ValueVectorEncoder.primitive[A, V], ValueVectorDecoder[V, A])
+//   implicit def primitive[A, V <: ValueVector](implicit schema: Schema[A]): ValueVectorCodec[A, V] =
+//     ValueVectorCodec(ValueVectorEncoder.primitive[A, V], ValueVectorDecoder[V, A])
 
-  implicit def list[A](implicit schema: Schema[A]): ValueVectorCodec[Chunk[A], ListVector] =
-    ValueVectorCodec(ValueVectorEncoder.list[A], ValueVectorDecoder.list[A])
+//   implicit def list[A](implicit schema: Schema[A]): ValueVectorCodec[Chunk[A], ListVector] =
+//     ValueVectorCodec(ValueVectorEncoder.list[A], ValueVectorDecoder.list[A])
 
-  implicit def struct[A](implicit schema: Schema[A]): ValueVectorCodec[A, StructVector] =
-    ValueVectorCodec(ValueVectorEncoder.struct[A], ValueVectorDecoder.struct[A])
+//   implicit def struct[A](implicit schema: Schema[A]): ValueVectorCodec[A, StructVector] =
+//     ValueVectorCodec(ValueVectorEncoder.struct[A], ValueVectorDecoder.struct[A])
 
-}
+// }
