+++ modules/core/src/main/scala/me/mnedokushev/zio/apache/arrow/core/codec/VectorSchemaRootCodec.scala
@@ -3,7 +3,7 @@ package me.mnedokushev.zio.apache.arrow.core.codec
 import org.apache.arrow.memory.BufferAllocator
 import org.apache.arrow.vector.VectorSchemaRoot
 import zio._
-import zio.schema.Schema
+// import zio.schema.Schema
 
 final case class VectorSchemaRootCodec[A](
   encoder: VectorSchemaRootEncoder[A],
@@ -30,12 +30,12 @@ final case class VectorSchemaRootCodec[A](
 
 }
 
-object VectorSchemaRootCodec {
+// object VectorSchemaRootCodec {
 
-  def apply[A](implicit codec: VectorSchemaRootCodec[A]): VectorSchemaRootCodec[A] =
-    codec
+//   def apply[A](implicit codec: VectorSchemaRootCodec[A]): VectorSchemaRootCodec[A] =
+//     codec
 
-  implicit def schema[A: Schema: SchemaEncoder]: VectorSchemaRootCodec[A] =
-    VectorSchemaRootCodec(VectorSchemaRootEncoder.schema[A], VectorSchemaRootDecoder.schema[A])
+//   implicit def schema[A: Schema: SchemaEncoder]: VectorSchemaRootCodec[A] =
+//     VectorSchemaRootCodec(VectorSchemaRootEncoder.schema[A], VectorSchemaRootDecoder.schema[A])
 
-}
+// }
